chapter-8/code/consumer/requirements.txt
ibmcloudant >= 0.1.3
paho-mqtt
